requirements.txt
appdirs==1.4.4
arrow==1.1.1
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
black==21.7b0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-yasg==1.20.0
flake8==3.9.2
idna==3.2
inflection==0.5.1
iniconfig==1.1.1
isort==5.9.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==21.0
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
pytz==2021.1
regex==2021.7.6
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
stripe==2.60.0
toml==0.10.2
tomli==1.0.4
uritemplate==3.0.1
urllib3==1.26.6
